% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxFw_maxFwBw_and_almost_unique_tol.R
\name{almost_unique_tol}
\alias{almost_unique_tol}
\title{almost_unique_tol}
\usage{
almost_unique_tol(vect, tol)
}
\arguments{
\item{vect}{a vector}

\item{tol}{tolerance}
}
\value{
a vector
}
\description{
almost_unique_tol
}
\examples{
vect <- c(69, 70, 77, 81, 85, 110)
tol = 2
almost_unique_tol(vect, tol)
}
