% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmseg_gauss_homoscd.R
\name{bruteForce_gauss_homoscd}
\alias{bruteForce_gauss_homoscd}
\title{Title}
\usage{
bruteForce_gauss_homoscd(y, X, K)
}
\arguments{
\item{y}{a vector of n observations (supposed gaussian and homoscedastic)}

\item{X}{a n x p matrix of covariates. The first column of X is filled with 1 and always corresponds to the intercept. Thus, if p = 1, then there are no covariates}

\item{K}{a parameter specifying that we want K-1 breakpoints}
}
\value{
a list containing: the regression results and the computed breakpoint maximizing the log-likelihood
}
\description{
Title
}
\examples{
n = 1000 ; X = matrix(1,n,1)
bp = c(553)
par = c(10,12,3)
y <- simuData_gauss_homoscd(X,par,bp)
K <- 2
res_bruteForce_gauss_homoscd =  bruteForce_gauss_homoscd(y,X,K)
res_bruteForce_gauss_homoscd
res_bruteForce_gauss_homoscd$bpForceBrute
}
