% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmseg_gauss_homoscd.R
\name{ComputeCritL_gauss_homoscd}
\alias{ComputeCritL_gauss_homoscd}
\title{ComputeCritL_gauss_homoscd}
\usage{
ComputeCritL_gauss_homoscd(y, X, K, bp_ProbRange)
}
\arguments{
\item{y}{a vector of n observations (supposed gaussian and homoscedastic)}

\item{X}{a n x p matrix of covariates. The first column of X is filled with 1 and always corresponds to the intercept. Thus, if p = 1, then there are no covariates}

\item{K}{a parameter specifying that we want K-1 breakpoints}

\item{bp_ProbRange}{a range of values including a given breakpoint}
}
\value{
a list containing respectively: the log-likelihood values and the BIC values for each of the points in bp_ProbRange
}
\description{
ComputeCritL_gauss_homoscd
}
\examples{
n = 1000 ; X = matrix(1,n,1)
bp = c(553)
par = c(10,12,3)
y <- simuData_gauss_homoscd(X,par,bp)
K <- 2
largeRange <- round(unname(quantile(1:n, c(0.05,0.95))))
bp_ProbRange = largeRange[1]:largeRange[2]
ComputeCritL_gauss_homoscd(y,X,K,bp_ProbRange)
}
