% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmseg_gauss_homoscd.R
\name{estim_gauss_homoscd}
\alias{estim_gauss_homoscd}
\title{estim_gauss_homoscd}
\usage{
estim_gauss_homoscd(y, X, bp = NULL)
}
\arguments{
\item{y}{a vector of n observations (supposed gaussian and homoscedastic)}

\item{X}{a n x p matrix of covariates. The first column of X is filled with 1 and always corresponds to the intercept. Thus, if p = 1, then there are no covariates}

\item{bp}{a vector of K-1 fixed breakpoints}
}
\value{
a list containing: a vector beta of estimated regression parameters for each sub-segment, the estimated standard deviation sigma common to all sub-segments, and the estimated logLikelihood loglik corresponding to the sum of the log-likelihoods on all sub-segments
}
\description{
estim_gauss_homoscd
}
\examples{
n = 1000 ; X = matrix(1,n,1)
bp = c(553)
par = c(10,12,3)
y <- simuData_gauss_homoscd(X,par,bp)
res_estimated <- estim_gauss_homoscd(y,X,bp)
res_estimated
}
