% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxFw_maxFwBw_and_almost_unique_tol.R
\name{Fw_factLog_Niv2}
\alias{Fw_factLog_Niv2}
\title{Fw_factLog_Niv2}
\usage{
Fw_factLog_Niv2(pi_jk, le, FUN)
}
\arguments{
\item{pi_jk}{a KxK matrix specifying that we want K-1 breakpoints (to be simplified !!!)}

\item{le}{a n*K matrix corresponding to the logarithm of the emission probabilities}

\item{FUN}{an operator allowing to choose between Forward (FUN = sum) and MaxForward (FUN = max) algorithms. We use MaxForward algorithm here}
}
\value{
log-likelihood
}
\description{
Fw_factLog_Niv2
}
