% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxFw_maxFwBw_and_almost_unique_tol.R
\name{FwBw_factLog_Niv2}
\alias{FwBw_factLog_Niv2}
\title{FwBw_factLog_Niv2}
\usage{
FwBw_factLog_Niv2(pi_jk, le, FUN, algo)
}
\arguments{
\item{pi_jk}{a KxK matrix specifying that we want K-1 breakpoints (to be simplified !!!)}

\item{le}{a n*K matrix corresponding to the logarithm of the emission probabilities}

\item{FUN}{an operator allowing to choose between Forward (FUN = sum) and MaxForward (FUN = max) algorithms. We use MaxForward algorithm here}

\item{algo}{a parameter allowing to choose between Forward-Backward (algo = 1) and MaxForward-MaxBackward (algo = 2) algorithms. We use MaxForward-MaxBackward algorithm here}
}
\value{
log-likelihood, marginal probabilities and segmentation
}
\description{
FwBw_factLog_Niv2
}
