% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfpopseg.R
\name{gfpop_gauss}
\alias{gfpop_gauss}
\title{gfpop_gauss}
\usage{
gfpop_gauss(y, K)
}
\arguments{
\item{y}{a vector of n observations (supposed gaussian and homoscedastic)}

\item{K}{a parameter specifying that we want K-1 breakpoints}
}
\value{
gfpop results and computed breakpoints
}
\description{
gfpop_gauss
}
