% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmseg_gauss_homoscd.R
\name{glmseg_gauss_homoscd}
\alias{glmseg_gauss_homoscd}
\title{glmseg_gauss_homoscd}
\usage{
glmseg_gauss_homoscd(y, X, K, ...)
}
\arguments{
\item{y}{a vector of n observations (supposed gaussian and homoscedastic)}

\item{X}{a n x p matrix of covariates. The first column of X is filled with 1 and always corresponds to the intercept. Thus, if p = 1, then there are no covariates}

\item{K}{a parameter specifying that we want K-1 breakpoints}

\item{...}{optional glm parameters (to be done completely !!!)}
}
\value{
list containing MaxEM results (log-likelihood and estimated parameters) and computed breakpiint(s)
}
\description{
glmseg_gauss_homoscd
}
\examples{
n = 1000 ; X = matrix(1,n,1)
bp = c(553)
par = c(10,12,3)
y <- simuData_gauss_homoscd(X,par,bp)
K = 2
glmseg_gauss_homoscd(y,X,K)
}
