% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxFw_maxFwBw_and_almost_unique_tol.R
\name{matPijk}
\alias{matPijk}
\title{matPijk}
\usage{
matPijk(K)
}
\arguments{
\item{K}{represents the number of segments (K-1 breakpoints)}
}
\value{
a KxK ("transition") matrix (in logarithmic scale)
}
\description{
matPijk
}
\examples{
matPijk(4)
}
