% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmseg_gauss_homoscd.R
\name{max_crit_gauss_homoscd}
\alias{max_crit_gauss_homoscd}
\title{max_crit_gauss_homoscd}
\usage{
max_crit_gauss_homoscd(y, X, par, lpi_jk)
}
\arguments{
\item{y}{a vector of n observations (supposed gaussian and homoscedastic)}

\item{X}{a n x p matrix of covariates. The first column of X is filled with 1 and always corresponds to the intercept. Thus, if p = 1, then there are no covariates}

\item{par}{a vector of p*K+1 parameters: p parameters for each of the K sub-segments + 1 parameter corresponding to the standard deviation common to all sub-segments}

\item{lpi_jk}{a KxK matrix specifying that we want K-1 breakpoints (to be simplified !!!)}
}
\value{
log-likelihood
}
\description{
max_crit_gauss_homoscd
}
\examples{
n = 1000 ; X = matrix(1,n,1)
bp = c(553)
par = c(10,12,3)
y <- simuData_gauss_homoscd(X,par,bp)
pi_jk = matrix(c(1,0,1,1),nrow = 2) ; lpi_jk = log(pi_jk)
max_crit_gauss_homoscd(y,X,par,lpi_jk)
}
