% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfpopseg.R
\name{myGraphFUN}
\alias{myGraphFUN}
\title{myGraphFUN}
\usage{
myGraphFUN(do.this)
}
\arguments{
\item{do.this}{a string of the form Tj where j = 1, ...., 10 represents the number of breakpoints}
}
\value{
a graph needed to compute j breakpoints using gfpop algorithm
}
\description{
myGraphFUN
}
