% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmseg_gauss_homoscd.R
\name{simuData_gauss_homoscd}
\alias{simuData_gauss_homoscd}
\title{simuData_gauss_homoscd}
\usage{
simuData_gauss_homoscd(X, par, bp = NULL)
}
\arguments{
\item{X}{a n x p matrix of covariates. The first column of X is filled with 1 and always corresponds to the intercept. Thus, if p = 1, then there are no covariates}

\item{par}{a vector of p*K+1 parameters: p parameters for each of the K sub-segments + 1 parameter corresponding to the standard deviation common to all sub-segments}

\item{bp}{a vector of K-1 fixed breakpoints}
}
\value{
a vector of n (homoscedastic) gaussian observations
}
\description{
simuData_gauss_homoscd
}
\examples{
n = 1000 ; X = matrix(1,n,1)
bp = c(553)
par = c(10,12,3)
y <- simuData_gauss_homoscd(X,par,bp)
y
}
